requirements.txt
# Core dependencies
flask>=2.0.0
flask-cors>=4.0.0
pydantic>=2.7.1
python-dotenv>=1.0.0
uvicorn>=0.27.0

# PDF and Document Processing
reportlab>=4.2.2

# NLP and Text Processing
beautifulsoup4>=4.12.3

# YAML Handling
ruamel.yaml>=0.18.6

# LangChain and AI
langchain>=0.2.11
langchain-openai>=0.1.6
langchain-core>=0.2.36
langchain-community>=0.2.10
langchain-google-genai>=0.0.6

# Testing
pytest>=8.2.2

# Utilities
requests>=2.32.3
python-dateutil>=2.8.0
free-proxy>=1.1.1
Jinja2>=3.1.4

# Additional dependencies for machine learning and embeddings
torch==2.1.0
transformers==4.36.0
scikit-learn==1.3.0
numpy==1.24.3
sentence-transformers==2.2.2